// src/app/edit/data-types/types.ts
export type BookAlbumPageProps = {
  params: {
    id: string;
  };
};

export type PageContent = {
  id: number;
  frontContent: React.ReactNode;
  backContent: React.ReactNode | null;
};

export type AlbumBook3DProps = {
  name: string;
  pages: PageContent[];
  width?: number;
  height?: number;
  photoHeight?: number;
  photoWidth?: number;
  coverColor?: string;
  coverImage?: string | null;
  lastPhotoUrl?: string;
  lastPhoto?: ImageDataProps;
};

export type AlbumDataProps = {
  webSizePx?: string;
  webPhotoSizePx?: string;
  layoutPage?: string;
  bookName?: string;
  images?: Array<{
    id: string;
    filename: string;
    mimeType: string;
    s3Url: string;
    previewUrl: string;
    imageUrl: string;
    height: number;
    width: number;
    metadata?: {
      captureDate?: string;
      eventGroup?: string;
      isCover?: boolean;
      textAnnotation?: TextAnnotation;
      rotation?: number | null;
      zoom?: number | null;
      caption?: string;
      locationName?: string | null;
      event_tags?: string[] | string;
      zoomPosition?: ZoomPosition | null;
    };
  }>;
  bookDesign: string | string[];
};